# docs/api/openapi.yaml
openapi: 3.0.0
info:
  contact:
    email: mandarine.app@yandex.ru
    name: Mandarine Support
  description: API for web and mobile application Mandarine
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  title: Mandarine API
  version: 0.0.0
paths:
...
  /v0/geocode/forward:
    get:
      description: Request for geocoding. User must be logged in. In response will be
        returned coordinates.
      operationId: Geocode
      parameters:
        - in: query
          name: address
          required: true
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
      responses:
        "200":
          description: Geocoded coordinates
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.GeocodingOutput"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.ErrorOutput"
        "401":
          description: Unauthorized error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.ErrorOutput"
        "403":
          description: User is blocked or deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.ErrorOutput"
        "503":
          description: Geocoding service is unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.ErrorOutput"
      security:
        - BearerAuth: []
      summary: Geocode
      tags:
        - Geocoding API
  /v0/geocode/reverse:
    get:
      description: Request for reverse geocoding. User must be logged in. In response
        will be returned address.
      operationId: ReverseGeocode
      parameters:
        - in: query
          name: lat
          required: true
          schema:
            type: number
            format: decimal
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
        - in: query
          name: lng
          required: true
          schema:
            type: number
            format: decimal
      responses:
        "200":
          description: Reverse geocoded address
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.ReverseGeocodingOutput"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.ErrorOutput"
        "401":
          description: Unauthorized error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.ErrorOutput"
        "403":
          description: User is blocked or deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.ErrorOutput"
        "503":
          description: Geocoding service is unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.ErrorOutput"
      security:
        - BearerAuth: []
      summary: Reverse geocode
      tags:
        - Geocoding API
...
servers:
  - url: //localhost:8080
components:
  securitySchemes:
    BearerAuth:
      in: header
      name: Authorization
      type: apiKey
  schemas:
...
    v0.AddressOutput:
      properties:
        city:
          type: string
        country:
          type: string
        countryCode:
          type: string
        county:
          type: string
        formattedAddress:
          type: string
        houseNumber:
          type: string
        postcode:
          type: string
        state:
          type: string
        stateCode:
          type: string
        stateDistrict:
          type: string
        street:
          type: string
        suburb:
          type: string
      type: object
...
    v0.ErrorOutput:
      properties:
        message:
          type: string
        path:
          example: /api/v0/example
          format: url_path
          type: string
        status:
          maximum: 599
          minimum: 400
          type: integer
        timestamp:
          type: string
      required:
        - message
        - path
        - status
        - timestamp
      type: object
    v0.GeocodingOutput:
      properties:
        count:
          type: integer
        data:
          items:
            $ref: "#/components/schemas/v0.PointOutput"
          type: array
      type: object
...
    v0.PointOutput:
      properties:
        latitude:
          type: number
        longitude:
          type: number
      type: object
...
    v0.ReverseGeocodingOutput:
      properties:
        count:
          type: integer
        data:
          items:
            $ref: "#/components/schemas/v0.AddressOutput"
          type: array
      type: object
...